`to_s` all the tags
During a migration from 1.6 to 3.0 we discovered that the ability to use
symbols as tags is no longer available and raises a `TypeError: can't
dup Symbol` exception. The reason for this is that we are running each
of the tags through a block and calling `#dup` on each value without
checking what data type it is. It's a fair assumption that all values
_should_ be strings however our application has shown that this isn't
the case and we the DataDog collector handles this fine in 1.6 so I
would expect the same to be true in 3.0 unless explicitly stated
otherwise.

To address the issue here I'm running each value via `#to_s` before we
pass it to `#dup` which will ensure that it never gets a symbol and
chokes.

Closes #50.