Add a "bypass_sampling" option to metric methods
It is sometimes desirable to make a decision about sampling before
emitting the sampled metric, for instance if computing the set of tags
to associate with the metric is expensive. Even in these circumstances,
though, the metric sent to statsd should still be tagged with a sample
rate so that counts can be computed correctly.

This adds a "bypass_sampling" option. When true, metrics are always sent
to statsd, but are reported as having been sampled at the sample_rate
that is passed in. It's the caller's responsibility to ensure that
sampling actually occurs.