Truncate long event messages by default
 # What
Instead of raising a RuntimeError and dropping the event when its
formatted string version is greater than 8kb, make an effort to truncate
the message portion of the event down to a sendable size.

Add the option to disable this feature by passing in :truncate => false
to @statsd.event.

 # Why
It's not acceptable to have a silent failure around a seemingly
arbitrary message length limitation, especially with Datadog's event
monitors being so feature poor. This would be less of a problem if you
could compose a monitor to query "has there been an error in the last 5
minutes, or has there not been a success in the last hour", but as it
currently stands silent failures are a big weakness for our Datadog
integration.

The C# client library
[accepted](https://github.com/DataDog/dogstatsd-csharp-client/pull/55) a
similar change, so it should fly in the ruby CL as well.