Small optimizations (#139)
* Eliminate runtime check of constant

We can know when the file is required whether or not
`Process.clock_gettime` is supported.  At that time, we can define a
method that calls the appropriate time handler depending on what is
supported by the current Ruby.  This reduces the overhead of the `time`
method because it doesn't have to check whether Process supports
monotonic time on every call

* Use Process.clock_gettime for telemetry

We can avoid allocating time objects when we need to test if it's time
to send telemetry