Always append global tags, regardless if `:tags` option specified
Upgrading from 4.5.0 directly to 4.8.0 we noticed our simple `#event` invocation (EMPTY_OPTIONS) was not sending our global tags. With this PR we removed the guard that a `:tags` option must be present for the tags serializer to return its tags. We must call it regardless to make sure global tags are appended.