Silence a deprecation in Ruby 2.7
Currently in `Ruby 2.7` this error will show on serialization:

```
warning: instance variable @global_tags_formatted not initialized
```

This PR modifies the assignment so that instead of not happening if the
conditional is false, it assigns `nil` back to the ivar.

---

Note: If you'd like to see the warning, just set `$VERBOSE = true` in
`spec/spec_helper.rb`

Thank you!