Call MessageBuffer#close in MessageBuffer#reset
Currently, MessageBuffer#reset is called only when the process
is forked and the thread "DogStatsD Sender" dies. In such a case,
the thread "DogStatsD MessageBuffer" also dies, but the method
should recreate the thread whether it is alive or not because
the word "reset" literally means making it in the initial state.