Add a pre_sampling option to metrics methods
This allows sampling decisions to be made outside of metrics emission
but still have the sampling rate sent to the collector. If pre_sampling
is false/not set the existing behavior of sampling in the library is
used. When pre_sampling is true metrics are always sent regardless of
the sample rate. In this mode it is up to the caller to handle sampling.